requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.35.0
decord>=0.6.0
einops>=0.7.0
scikit-learn>=1.3.0
numpy>=1.24.0
opencv-python>=4.8.0
matplotlib>=3.7.0
tensorboard>=2.14.0
PyYAML>=6.0
tqdm>=4.65.0
timm>=0.9.0
